Code structure changes to seperate classes into seperate scripts and minor refactors to ammend PR comments
