+++ plotting/plot_factory/tests/test_plot_factory.py
@@ -14,15 +14,14 @@ Unit tests for plot factory
 #  - Add hard coded string equivalent of dictionary to test_dict_to_string
 #  - Add Error bars to yaml file options
 
+# Core Dependencies
 import unittest
-from mock import patch, Mock, PropertyMock
+from mock import patch, PropertyMock
 
-import pandas as pd
-import plotly
 import dash
 
 # Internal Dependencies
-from plotting.plot_factory.plot_factory import PlotFactory, Layout, Trace, DashApp
+from plotting.plot_factory.plot_factory import PlotFactory
 
 # Mocked values
 from plotting.plot_factory.tests.mocked_values import MockPlotVariables
@@ -34,7 +33,7 @@ class TestPlotFactory(unittest.TestCase):
     @patch('plotting.plot_factory.plot_factory.Layout')
     def test_get_trace_list(self, mock_layout, _):
         """
-        Test: get_trace_list is called within TestPlotFactory().construct_plot()
+        Test: create_trace_list is called within TestPlotFactory().create_plot()
         When: trace_list = []
         """
         mock_layout.return_value.mode = PropertyMock("mock_mode")
@@ -42,129 +41,23 @@ class TestPlotFactory(unittest.TestCase):
         layout = mock_layout
 
         # Assert that a list of trace object is returned
-        actual = PlotFactory().get_trace_list(
+        actual = PlotFactory().create_trace_list(
             MockPlotVariables().indexed_single_multi_raw_data_dataframe,
             layout, MockPlotVariables().plot_name)
         self.assertIsInstance(actual, list)  # is list
         self.assertEqual(len(actual), 2)
 
     # @patch('plotting.plot_factory.plot_factory.DashApp')
-    @patch('plotting.plot_factory.plot_factory.PlotFactory.get_trace_list')
+    @patch('plotting.plot_factory.plot_factory.PlotFactory.create_trace_list')
     @patch('plotting.plot_factory.plot_factory.Layout')
     def test_construct_plot(self, mock_layout, mock_get_trace):
         """
         Test: dashapp object is returned
-        When: called with self.get_trace_list() and Layout().layout
+        When: called with self.create_trace_list() and Layout().layout
         """
-        actual = PlotFactory().construct_plot(
+        actual = PlotFactory().create_plot(
             plot_meta_file_location=MockPlotVariables().plot_meta_file_location,
             data=MockPlotVariables().raw_multi_single_data_dataframe,
             figure_name=MockPlotVariables().plot_name)
 
         self.assertIsInstance(actual.app, dash.dash.Dash)
-
-
-class TestLayout(unittest.TestCase):
-
-    @patch('plotting.plot_meta_language.interpreter.Interpreter.interpret')
-    def test_read_plot_meta_data(self, mock_interpreter):
-        """
-        Test: read_plot_meta_data is called within Layout()
-        When: Layout().extract_layout() is called with meta_file_location
-        """
-        mock_interpreter.return_value = MockPlotVariables().raw_interpreted_layout
-
-        actual = Layout(MockPlotVariables().plot_meta_file_location).read_plot_meta_data()
-
-        # assert that interpreter has been called with correct argument
-        mock_interpreter.assert_called_with(MockPlotVariables().plot_meta_file_location)
-
-        # return value assertions
-        self.assertIsInstance(actual, dict)
-        self.assertEqual(actual, MockPlotVariables().raw_interpreted_layout)
-
-    @patch('plotting.plot_factory.plot_factory.Layout.read_plot_meta_data')
-    def test_extract_layout(self, mock_rpmd):
-        """
-        Test: extract_layout is called within Layout() returning layout values in dictionary only
-        When: mode and plot values exist in dict returned by read_plot_meta_data()
-        """
-
-        mock_rpmd.return_value = MockPlotVariables().raw_interpreted_layout
-
-        actual = Layout(
-            MockPlotVariables().raw_interpreted_layout).extract_layout()
-
-        self.assertIsInstance(actual, dict)
-        self.assertEqual(actual, MockPlotVariables().processed_layout)
-
-
-class TestTrace(unittest.TestCase):
-
-    def setUp(self):
-        """Contains setup variables to be used in unit tests for Trace class methods"""
-
-        self.error_bars = True
-
-    def test_trace_dict__error_bars_set_true(self):
-        """
-        Test: trace_dict() returns a dict in expected trace format containing error_bars
-        When: called with dataframe and error_bars=True
-        """
-
-        actual = MockPlotVariables().trace_object.trace_dict(
-            data=MockPlotVariables().indexed_multi_single_raw_data_dataframe,
-            error_bars=self.error_bars)
-
-        self.assertIsInstance(actual, dict)
-        self.assertEqual(actual, MockPlotVariables().trace_multi_single)
-
-    def test_trace_dict__error_bars_set_false(self):
-        """
-        Test: trace_dict() returns a dict in expected trace format without error_bars
-        When: called with dataframe and error_bars=False
-        """
-        actual = MockPlotVariables().trace_object.trace_dict(
-            MockPlotVariables().indexed_multi_single_raw_data_dataframe, False)
-
-        self.assertIsInstance(actual, dict)
-        self.assertEqual(actual, MockPlotVariables().trace_multi_single_error_y_not_visible)
-
-    def test_dict_to_string(self):
-        """
-        Test: dict_to_string() is returned as string in expected format
-        When: called with dictionary by create_trace()
-        """
-
-        actual = MockPlotVariables().trace_object.dict_to_string(
-            MockPlotVariables().trace_multi_single_to_string)
-
-        self.assertIsInstance(actual, str)
-
-    def test_create_trace(self):
-        """
-        Test: _create_connection() is called within connect()
-        When: client._connection is None
-        """
-
-        actual = MockPlotVariables().trace_object.trace
-
-        self.assertIsInstance(actual, plotly.graph_objs.Scattergl)
-
-
-class TestDashApp(unittest.TestCase):
-
-    def setUp(self):
-        """Contains setup variables to be used in unit tests for Trace class methods"""
-        self.figure = None
-
-    def test_create_dashapp(self):
-        """
-        Test:create_dashapp() is called returning an instance of DashApp object
-        When: called with figure and app_id by DashApp()
-        """
-        # Assert a DashApp is returned
-        actual = DashApp(self.figure, MockPlotVariables().plot_name).create_dashapp()
-
-        self.assertIsInstance(actual, dash.dash.Dash)
-
